Added source code maps to coffee script compilation
This is so we can see where errors are actually being reported.

I've also had to remove the require js dependency, though we weren't
using it any. But since I was unsure on how to get requirejs to generate
sourcemaps while taking existing source maps into consideration, I thought
it was prudent to just remove it.